Incorporated the Nashville meetup dataset.
Nashville meetup dataset is a dataset with a set of members that went to group meetups of a wide array of groups. There is a file of member interactions (which is used as edges) and a file of group attributes and members belonging to groups. The groups are divided into categories and by one hot encoding which members are part of groups of certain categories we create attributes for members with a usable distribution.