feat(TwitchDataset): Implement ablation study
As far as my understanding goes the idea of the ablation study is to fix the reference group to the entire dataset, meaning instead of discovering rho dynamically by discovery, we force it to be `rho = len(ranks)`.

This code adds the ablation_mode parameter which by default is false. If it is set to true, the rho is fixed to len(ranks) to calculate the ablation study